refactor: fab, modal section 분리
