+++ features/record-detail/sections/detail-cheer-fab.tsx
@@ -2,25 +2,17 @@
 
 import { useState } from 'react';
 
-import { useBottomSheet, useDragScroll, useModal } from '@/hooks';
+import { useBottomSheet } from '@/hooks';
 import { css } from '@/styled-system/css';
-import { flex } from '@/styled-system/patterns';
 
 import { useCheer, useCheerPreviewList } from '../apis';
-import {
-  CheerBottomSheet,
-  CheerItem,
-  CheerModal,
-  CheerProgress,
-} from '../components';
+import { CheerBottomSheet, CheerProgress } from '../components';
 import { initialCheerList } from '../data';
 import { DetailCheerItemSelected, RecordDetailType } from '../types';
 
-export const DetailCheer = ({ data }: { data: RecordDetailType }) => {
+export const DetailCheerFabSection = ({ data }: { data: RecordDetailType }) => {
   const { mutate: mutateCheer } = useCheer();
-  const { data: cheerPreviewData, refetch: refetchCheer } = useCheerPreviewList(
-    data.id,
-  );
+  const { refetch: refetchCheer } = useCheerPreviewList(data.id);
 
   const [cheerList, setCheerList] = useState(initialCheerList);
   const [selectedCheerItem, setSelectedCheerItem] =
@@ -31,13 +23,6 @@ export const DetailCheer = ({ data }: { data: RecordDetailType }) => {
     open: openBottomSheet,
     close: closeBottomSheet,
   } = useBottomSheet();
-  const {
-    isOpen: isOpenModal,
-    open: openModal,
-    close: closeModal,
-  } = useModal();
-
-  const { sliderRef } = useDragScroll();
 
   const handleClickCheerItem = (index: number) => {
     setCheerList((prev) =>
@@ -81,30 +66,9 @@ export const DetailCheer = ({ data }: { data: RecordDetailType }) => {
   };
 
   const { isMyMemory } = data;
-  const reactions = cheerPreviewData?.reactions || [];
+
   return (
     <>
-      {/* NOTE: 응원 미리보기 목록 */}
-      {reactions && Boolean(reactions.length) && (
-        <div className={slider.containerStyle} ref={sliderRef}>
-          <div className={slider.wrapperStyle}>
-            {reactions.map((item) => (
-              <CheerItem
-                {...item}
-                key={item.reactionId}
-                onClick={openModal}
-                size="small"
-              />
-            ))}
-            {reactions.length > 10 && (
-              <button className={slider.entireCheerButton} onClick={openModal}>
-                응원 전체보기
-              </button>
-            )}
-          </div>
-        </div>
-      )}
-
       {/* NOTE: 응원 FAB Button */}
       {!isMyMemory && (
         <button className={cheerButtonWrapperStyle} onClick={openBottomSheet}>
@@ -120,14 +84,6 @@ export const DetailCheer = ({ data }: { data: RecordDetailType }) => {
         cheerItem={selectedCheerItem}
       />
 
-      {/* NOTE: 응원 모달 */}
-      <CheerModal
-        isOpen={isOpenModal}
-        onClose={closeModal}
-        title="8월 16일의 응원"
-        description="5"
-      />
-
       {/* NOTE: 응원 바텀시트 */}
       <CheerBottomSheet
         header={{ title: '응원 보내기' }}
@@ -139,33 +95,6 @@ export const DetailCheer = ({ data }: { data: RecordDetailType }) => {
       />
     </>
   );
-};
-
-const slider = {
-  containerStyle: css({
-    overflowX: 'scroll',
-    backgroundColor: 'white',
-
-    '&::-webkit-scrollbar': {
-      display: 'none',
-    },
-  }),
-
-  wrapperStyle: flex({
-    gap: '10px',
-    width: 'max-content',
-    backgroundColor: 'white',
-    p: '20px 20px 0px',
-  }),
-
-  entireCheerButton: css({
-    textStyle: 'label1.normal',
-    fontWeight: 'medium',
-    color: 'background.white',
-    backgroundColor: 'text.neutral',
-    rounded: '7px',
-    p: '7px 12px',
-  }),
 };
 
 const cheerButtonWrapperStyle = css({
