+++ features/record-detail/sections/index.ts
@@ -1,4 +1,5 @@
-export * from './detail-cheer';
+export * from './detail-cheer-fab';
+export * from './detail-cheer-modal';
 export * from './detail-description';
 export * from './detail-diary';
 export * from './detail-preview';
